+++ src/pages/report/ProductReport/ProductSaleReport/index.tsx
@@ -9,122 +9,16 @@ import ReportDatePicker from 'pages/report/components/ReportDatePicker';
 import ReportPage from 'pages/report/components/ReportPage';
 import { useEffect, useRef, useState } from 'react';
 import { useForm } from 'react-hook-form';
-// components
 import { useNavigate } from 'react-router-dom';
-//
-import { TTableColumn } from 'types/table';
-import { fNumber, fPercent } from 'utils/formatNumber';
+import productSaleColumn from './column';
 
 const ProductSaleReport = () => {
   const navigate = useNavigate();
   const { translate } = useLocales();
   const { enqueueSnackbar } = useSnackbar();
+  const { watch } = useForm();
 
-  type ProductSaleDetail = {
-    storeId?: any;
-    productId?: any;
-    productName?: any;
-    quantity?: any;
-    percent?: any;
-    totalBeforeDiscount?: any;
-    discount?: any;
-    totalAfterDiscount?: any;
-    checkDeal?: any;
-  };
-
-  const orderColumns: TTableColumn<ProductSaleDetail>[] = [
-    {
-      title: 'Cửa hàng',
-      hideInTable: true,
-      valueType: 'select',
-      dataIndex: 'storeId'
-    },
-    {
-      title: 'Ngày',
-      hideInTable: true,
-      valueType: 'select',
-      hideInSearch: true
-    },
-    {
-      title: 'Tên sản phẩm',
-      hideInSearch: true,
-      dataIndex: 'productName'
-    },
-    {
-      title: 'Số lượng',
-      hideInSearch: true,
-      dataIndex: 'quantity',
-      render: (value) => fNumber(value)
-    },
-    {
-      title: 'Tỉ trọng (%)',
-      hideInSearch: true,
-      dataIndex: 'percent',
-      render: (value) => fPercent(value)
-    },
-    {
-      title: 'Tổng tiền trước giảm giá',
-      hideInSearch: true,
-      dataIndex: 'totalBeforeDiscount',
-      render: (value) =>
-        value.toLocaleString('vi', {
-          style: 'currency',
-          currency: 'VND'
-        })
-    },
-    {
-      title: 'Giảm giá',
-      hideInSearch: true,
-      dataIndex: 'discount',
-      render: (value) =>
-        value.toLocaleString('vi', {
-          style: 'currency',
-          currency: 'VND'
-        })
-    },
-    {
-      title: 'Tổng tiền sau giảm giá',
-      hideInSearch: true,
-      dataIndex: 'totalAfterDiscount',
-      render: (value) =>
-        value.toLocaleString('vi', {
-          style: 'currency',
-          currency: 'VND'
-        })
-    },
-    {
-      title: 'Loại',
-      hideInTable: true,
-
-      valueEnum: [
-        {
-          label: 'Sản phẩm',
-          value: 'true'
-        },
-        {
-          label: 'Nhóm sản phẩm',
-          value: 'false'
-        }
-      ],
-      valueType: 'select'
-    },
-    {
-      title: 'Chọn biểu đồ',
-      hideInTable: true,
-      dataIndex: 'checkDeal',
-      valueEnum: [
-        {
-          label: 'Giảm giá',
-          value: 'true'
-        },
-        {
-          label: 'Trước giảm giá',
-          value: 'false'
-        }
-      ],
-      valueType: 'select'
-    }
-  ];
+  const ref = useRef<any>();
 
   const [openChart, setOpenChart] = useState(false);
   const today = new Date();
@@ -132,8 +26,6 @@ const ProductSaleReport = () => {
   const [fromDate, setFromDate] = useState<Date>(new Date(yesterday));
   const [toDate, setToDate] = useState<Date>(new Date());
 
-  const ref = useRef<any>();
-  const { watch } = useForm();
   useEffect(() => {
     if (ref.current) {
       ref.current.formControl.setValue(
@@ -198,7 +90,7 @@ const ProductSaleReport = () => {
 
           <ResoTable
             ref={ref}
-            columns={orderColumns}
+            columns={productSaleColumn}
             getData={productApi.getProductReport}
             showAction={false}
             pagination={true}