+++ src/types/report/productProgress.ts
@@ -1,0 +1,7 @@
+export type ProductProgressBase = {
+  date?: any;
+  quantity?: any;
+  revenueBefore?: any;
+  discount?: any;
+  revenue?: any;
+};