+++ src/types/report/productSale.ts
@@ -1,0 +1,11 @@
+export type ProductSaleBase = {
+  storeId?: any;
+  productId?: any;
+  productName?: any;
+  quantity?: any;
+  percent?: any;
+  totalBeforeDiscount?: any;
+  discount?: any;
+  totalAfterDiscount?: any;
+  checkDeal?: any;
+};